refactor(decomp): extract ExecutionResult → query-fusion (TD-DECOMP-32) STRUCTURAL CASCADE
Move FederatedExecutionResult + FederatedExecutionStats + ExecutionResult type alias
into query-fusion crate (new execution_result.rs module). Root execution/mod.rs
re-exports them. This unblocks query_result_cache (19t) + invalidation (11t) = 30
tests for the next cascade step into query-cache.

The types are clean data containers (Arrow + DataModel only, all pub methods).
ModelType = proximadb_data_model::DataModel (foundation crate, already a dep).
query-fusion gains arrow.workspace = true.

STRUCTURAL CASCADE step 1 of the TD-DECOMP-29 follow-up. Verified: standalone
query-fusion --all-targets EXIT 0 + ROOT LIB CHECK zero errors (BOTH on first try
— the pre-extraction checklist works). See TD-DECOMP-32.
