+++ crates/query/proximadb-query-fusion/Cargo.toml
@@ -18,3 +18,4 @@ proximadb-data-model = { path = "../../foundation/proximadb-data-model" }
 serde.workspace = true
 serde_json.workspace = true
 anyhow.workspace = true
+arrow.workspace = true
