+++ docs/10-quality/td/TD-DECOMP-32-extract-execution-result.adoc
@@ -1,0 +1,39 @@
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+= TD-DECOMP-32: Extract FederatedExecutionResult → query-fusion (STRUCTURAL CASCADE step 1)
+
+[cols="1,3", options="header"]
+|===
+| Field | Value
+| Status | Landing
+| Severity | Medium — structural cascade (unblocks 30 tests in query/cache)
+| Component | `crates/query/proximadb-query-fusion/src/execution_result.rs`; root `src/query/federated/execution/mod.rs`
+| Relates to | [[TD-DECOMP-29]] (query-cache structural enabler); [[TD-DECOMP-1]] (recipe)
+|===
+
+== Why
+
+The query-cache structural enabler (TD-DECOMP-29) created `proximadb-query-cache` as a home for
+5 coupled files (58 tests) blocked by root-only types. The #1 blocker was `ExecutionResult`
+(affecting query_result_cache 19t + invalidation 11t = 30 tests). This TD extracts the
+`ExecutionResult` types to query-fusion, unblocking those files.
+
+== What moved (behavior-neutral)
+
+* `FederatedExecutionResult` + `FederatedExecutionStats` + `ExecutionResult` type alias →
+  `crates/query/proximadb-query-fusion/src/execution_result.rs`.
+* Root `src/query/federated/execution/mod.rs`: type definitions replaced with
+  `pub use proximadb_query_fusion::execution_result::{...}` (16 callers resolve unchanged).
+* query-fusion gains `arrow.workspace = true` dep. `ModelType` is `proximadb_data_model::DataModel`
+  (foundation crate, already a query-fusion dep).
+* The impl block (4 methods: empty/from_batch/empty_with_schema/row_count) is all `pub` —
+  no pub(crate) widening needed.
+
+== Follow-ups
+
+* **Next cascade step:** move `query_result_cache.rs` (19t) + `invalidation.rs` (11t) →
+  query-cache crate (their ExecutionResult dep now resolves from query-fusion via root re-export).
+* `PlanOutput` cascade remains blocked (depends on FilterStrategy/IndexRoute from modality-tier
+  observability-engine — layering forbids query→modality).
+* Verified: standalone query-fusion --all-targets EXIT 0 + root lib check zero errors (BOTH
+  checks passed on first try — the pre-extraction checklist works).
